Use pyflakes' api rather than shelling out
We've been shelling out to pyflakes and parsing its output. This is all well
and good except that it's... well, silly. Pyflakes is a python library, and
provides us with well structured data faster if we use it like so. This halves
the runtime of our unit tests (from 6 to 3 seconds on my system)!