Only check for extra tor processes once
Well, that was unintuitive. Our test class' constructor is called once for each
test, so self.is_extra_tor_running was always None for our setup block.

Moving this out into a global so we only check this once. This has a couple
advantages...

1. Fewer system calls so our integ tests run faster. This amounts to about 0.5s
   on my system.

2. Less confusing test output. Our jenkins tests had a handful of failures that
   was probably due to having multiple tor processes running...

   https://jenkins.torproject.org/job/stem-tor-ci/448/console

   The weird thing was that some tests properly skipped with 'multiple tor
   instances' while others didn't and in turn failed. Now we should either
   always skip or always run.

One thing that has me concerned is that if a new tor instance spins up in the
middle of our test run then we won't detect it. We might need to change this
later to always check if we last detected only a single process. Even if we do
this we'll get most of the runtime benefits since this now resides in a helper
function rather than setup (so it's only triggered when relevant).