Don't use @lru_cache for exit policy rules
Lession learned today: @lru_cache is handy for efficiency but don't rely on the
cache to function.

Our ExitPolicy class expected its _get_rules() method to be called exactly once
due to having cached results. However, pickling then unpicking an ExitPolicy
class caused our @lru_cache to be cleared, resulting in a TypeError.

Replacing this method's @lru_cache with a standard one, and adding a test for
pickleability. Thanks to Nicholas Hopper for the catch!