Better connect() function for easily getting a Controller
Stem provided a couple high level functions for getting a Controller with
minimal fuss: connect_port() and connect_socket_file(). These interacted
directly with stdin and stdout making them undesirable for sophisticated use
cases, but for commandline scripts provided a one-line method of geting an
authenticated Controller.

Personally I never used these myself. Not because I didn't need this, but
because I had done better in arm. In arm's connection helper it...

* Provided better, more useful error output upon authentication failure.

* Attempted to connect to both a control port and socket, then was smart about
  the message it presented if both failed.

On reflection the first belongs in Stem (hell, why not provide better error
output?!?). As for the second I realized that having separate control_port()
and control_sockeet_file() functions was a mistake.

This deprecates the old connect_* functions in favor of a single new connect()
that does essentially exactly what we do in arm: try to get a tor connection,
and give nice advice if we can't.