Connect didn't flag controllers as authenticated
Our Controller class tracks if it's been authenticated to Tor or not. When
users call msg('AUTHENTICATE...') we're authenticated, and when we attach to a
new socket we're unauthenticated.

This is all well and good, but when a user first constructs us from a
ControlSocket we have no idea if that socket has been authenticated or not. We
can't actively check because Tor will disconnect the socket if it doesn't obey
a very specific authentication dance (... ugg). So we need our caller to tell
us if it's been authenticated or not.

Previously we just assumed it wasn't authenticated, causing our connect(),
connect_port(), and connect_socket_file() methods to return a Controller that
has been authenticated, but didn't know it had. In general this worked, but
turns out event listening checks the authenticated flag.