Changing @uses_settings to use keyword argument
Our @uses_settings passed along the configuration as the function's first
argument. This is well and good for functions, but for methods that meant we no
longer started with 'self'. This is really, really weird in the python world so
changing the annotation to pass the value along as a keyword argument.