Autocompletion unit tests
Unit tests for our autocompletion class and helper funtion for determine
options from tor. This didn't uncover any issues with that, but it *did* reveal
a bug with our util module. We didn't make a shallow copy of list config
values, so in place modifications edited the configuration's master value.