Fix interaction of color prompt with readline
For a long while we've had an issue where readline history would break with
long inputs. For instance, entering...

  >>> the quick brown fox jumped over the very nice fence and ran around in pretty fields

... then pressing 'up' to autocomplete history then 'down' to go back to an
empty prompt would leave a fragment of this mingled with the prompt.

This turned out to be an issue with how readline calculates column width.
Readline included the formatting sequences when determining our prompts width,
which can be corrected by wrapping all encodings with RL_PROMPT_START_IGNORE
and RL_PROMPT_END_IGNORE sequences...

  https://stackoverflow.com/questions/9468435/look-how-to-fix-column-calculation-in-python-readline-if-use-color-prompt

Many thanks to Yawning for pointing this out!