Expanding /info output
Greatly enriching our /info output to include all relay information the user is
likely to find helpful. This revises the summary a bit, but more importantly
now includes the relay's hostname and all descriptor content (server,
extrainfo, microdescriptor, and router status entry).

Fetching both the hostname and descriptor content involves active lookups, but
while this was a no-go for arm it's perfectly fine for an interactive
interpreter like this.