Changing bridge descriptor link to CollecTor
Sanatized bridge descriptors are described on both Metrics' formats page and
CollecTor...

  https://metrics.torproject.org/formats.html#bridgedesc
  https://collector.torproject.org/formats.html#bridge-descriptors

Switching to the later. I believe the descriptor types are the only thing we
still link to on the old format page...

  https://metrics.torproject.org/formats.html#descriptortypes

However, it doesn't have a direct counterpart in CollecTor. It enumerates all
descriptor annotations, but CollecTor only includes the annotations it uses.