Adding crop() method to our str_tools
Method from arm's ui_tools with a little cleanup. This is a pretty delicate
method and I'm a tad worried about some changes so we might see some regression
from what arm had. That said, regressions can easily be added to our unit tests
(it'll be soooo nice to finally have those for this helper....).