Having create_hidden_service() return the *.onion address
This isn't reliable since tor restricts the hidden service directory to only be
readable by the tor user. This sucks, since it's handy to have the hidden
service address and this is the only way to get it. But oh well, more useful
than returning a boolean as we did before.