Dropping unnecessary list() calls
Python's 2to3 converter wraps quite a few things with list() to be safe...

  https://stackoverflow.com/questions/17695456/why-python-3-needs-wrap-dict-items-with-list

More often than not this is the right thing. In particular...

  * If we return the value then we need the list() call, since dict's keys(),
    values(), and items() now return a different type.

  * If we concurrently modify the dictionary *or* it's an argument provided
    from our caller (so they might concurrently modify it) then we need this
    shallow copy so we don't raise an exception due to concurrent modification.

That is to say, the only time we *don't* need the list() call is when it's a
local variable that won't be concurrently modified. There's quite a few
instances of this, so dropping the extraneous list() when we can.