Don't modify list of descriptor lines
In theory popping from the head of our unread lines is inefficient since we
create a new list with every line read. However, oddly changeing this doesn't
seem to have a benefit in terms of performance...

It's more readable how it was so including this commit and reversion in our
history so we can loop back and include this if we later discover it is
beneficial.