Use full tor path for pidof integ test
When running our tests with the '--tor' argument the first time we run the
pidof test it passes, but further runs fail. I'm not sure why that first one
passes, but in checking the later ones it's picking up our 'run_tests.py --tor
/path/to/tor' since it includes tor in it.

The pidof man page suggests providing an absolute path so avoiding truncation
of the command for this test.

  ======================================================================
  FAIL: test_pid_by_name_pidof
  ----------------------------------------------------------------------
  Traceback (most recent call last):
    File "/home/atagar/Desktop/stem/test/integ/util/system.py", line 145, in test_pid_by_name_pidof
      self.assertEqual(tor_pid, stem.util.system.pid_by_name(tor_cmd))
  AssertionError: 10020 != None

  ----------------------------------------------------------------------
  Ran 26 tests in 1.435s