Module for Tor manual information
Nyx, Erebus, and Stem's interpreter all have use for human readable information
about tor's configuration options. Nyx has long parsed the man page for this
but like most of Nyx it was pretty cobbled together. No tests, poor
performance, and not very general purpose.

  https://trac.torproject.org/projects/tor/ticket/8251

Moving the useful bits of Nyx over and expanding it to be an order of magnitude
faster (by calling 'man -P cat' to avoid its pager) and have a great compliment
of tests.

Note that the tests presently expect the tor man page in our test directory but
it doesn't exist. I'll be addressing this in an upcoming commit, but I don't
want to add the man page to our git history (no need to artificially inflait
our repository size).