Test that we install all files
Few times now we've had bugs (and even releases!) where installation misses
files. This is because in our setup.py we need to enumerate all modules and
non-source files. This is damn easy to forget.

Renaming stem/settings.cfg hit this again so adding a test to ensure it stops
happening in the future.