Cached tor manual information
Adding the Manual.from_cache() and save() methods. With this users can use
manual information bundled with Stem. This is faster, more reliable, and
doesn't require any system commands. Only drawback is that the information is
only as up to date as Stem's release.

This includes a 'cache_manual.py' script that synchronizes our manual
information with tor. Handy when the man page changes.