Truncate trace level logging
At the trace level we have request/reply logging. Very useful, but some tor
requests like 'GETINFO ns/all' are huge. Truncating responses to ten lines by
default. This can be customized via a constant.

This reduces the debug log when nyx is run for a few seconds from 1.6M to 80K,
mostly because it no longer causes us to dump the full consensus into it.