Revise download page's packages for Debian and Fedora
avet on irc expressed concern because our download page cites python-stem but
not python3-stem. Good thing to cite both. As for Fedora we cited a couple
derived packages that don't actually seem to exist. This is fine. Reached out
to Juan to confirm, but they didn't have much reason to exist so probably just
out of date.