Sort fallbacks by fingerprint
Updating our fallback directories caused a big delta, not because of the
changes but because we have a random ordering. Sorting our config file by the
fallback fingerprints so we have small deltas in the future.

Only functional change to the config file is the following to get us in sync...

  * Removed FCB6695F8F2DC240E974510A4B3A0F2B12AB5B64 as a fallback directory
  * Removed 14419131033443AE6E21DA82B0D307F7CAE42BDB as a fallback directory
  * Removed 5525D0429BFE5DC4F1B0E9DE47A4CFA169661E33 as a fallback directory
  * Removed 29F1020B94BE25E6BE1AD13E93CE19D2131B487C as a fallback directory
  * Removed E781F4EC69671B3F1864AE2753E0890351506329 as a fallback directory
  * Removed DE134FC8E5CC4EC8A5DE66934E70AC9D70267197 as a fallback directory
  * Removed 6A640018EABF3DA9BAD9321AA37C2C87BBE1F907 as a fallback directory
  * Removed 6FC6F08270D565BE89B7C819DD8E2D487397C073 as a fallback directory