Fill in missing manual summaries
As tor adds new config options we want short, particularly human-readable
summaries. Filling in the new additons.

We have a test that checks for this but it required the ONLINE integ target so
really never gets run in practice. Replacing this with a unit test since
honestly we don't need the downloaded information for this one.