Stem release 1.5.2
Release creation in PyPI is both error prone and unfixable. As such it's safest
if we have a scrap project to sanity check that things are working as intended
before permanently screwing up our package. Really silly we need to do this,
but that's life.

  https://github.com/pypa/packaging-problems/issues/74#issuecomment-260716129

Previously we made releases with 'python setup.py upload' but PyPI has broken
this by prohibiting non-SSL uploads. Twine now seems to be the only game in
town so updating our setup.py clauses with things we previously set via the
site. One caveat is that PyPI's bug tracker field evidently cannot be set
via the setup.py...

  https://stackoverflow.com/questions/14459828/how-to-set-bug-tracker-url-in-setup-py-script

Oh well. Close enough.