Stop checking consensus field ordering
Tor's spec details the order that consensus fields should appear in, but this
is a detail tor frequently gets wrong. The spec's been changed to loosen this
requirement...

  https://trac.torproject.org/projects/tor/ticket/21059
  https://gitweb.torproject.org/torspec.git/commit/?id=578477a

As such Stem should no longer validate this. New misordered additions in the
live tor consensus have been causing our validator to alarm but with this it's
once again happy.

Turns out the order check also masked some issues with our missing fields test.
Took a while to figure out but in-place modification of our 'attr' subtly broke
it. Addressing this with a defensive shallow copy.