Only require 'pycodestyle.ignore' snippets to only need the prefix
Changing our ignore rules to be a prefix match rather than an exact match. This
way users don't need to deal with the headache of inline comments and such.

We needed a new ignore rule because the version of pycodestyle differs between
my netbook and pc, so a minor issue slipped in.