Move key_certificate into validate_signature()
Does it work? We don't have a test so haven't a clue. But it's a whole lot
cleaner than trying to wedge a new attribute into our DirectoryAuthority
instances.

I can see why we previously did this since it allows the constructor to perform
the crypto check, but since this requires an extra ingredient (the signing
keys) it's better as either a separate validate function or a private
attribute.